Extract method into its own crate
