+++ rs/registry/canister/src/mutations/node_management/common.rs
@@ -222,7 +222,7 @@ pub fn node_exists_with_ipv4(registry: &Registry, ipv4_addr: &str) -> bool {
 
 /// Similar to `get_key_family` on the `RegistryClient`, return a list of
 /// tuples, (ID, value).
-fn get_key_family<T: prost::Message + Default>(
+pub(crate) fn get_key_family<T: prost::Message + Default>(
     registry: &Registry,
     prefix: &str,
 ) -> Vec<(String, T)> {
